+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2026-08-21]
+## [1.1.0 - 2026-08-24]
 ### Added
 - Dockerfile
 - Single Spa
@@ -133,6 +133,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - NOT SELECTED configuration
 - Unique feasibility statistics table
 - Bridgehead contact data
+- Multiple form fields
 
 ### Changed
 - Display feasibility results based only on non-empty backend results